# AIForRentersAPI/AIForRentersAPI/bin/Release/netcoreapp3.1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AIForRentersAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Availabilities": {
      "get": {
        "tags": [
          "Availabilities"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Availabilities"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Availability"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Availability"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Availability"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          }
        }
      }
    },
    "/api/Availabilities/{id}": {
      "get": {
        "tags": [
          "Availabilities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Availabilities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Availability"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Availability"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Availability"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Availabilities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          }
        }
      }
    },
    "/api/Clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      }
    },
    "/api/Clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Property": {
        "required": [
          "location",
          "name"
        ],
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "location": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "unit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Unit"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Unit": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "unitId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "propertyId": {
            "type": "integer",
            "format": "int32"
          },
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "availability": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Availability"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Availability": {
        "type": "object",
        "properties": {
          "availabilityId": {
            "type": "integer",
            "format": "int32"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time"
          },
          "toDate": {
            "type": "string",
            "format": "date-time"
          },
          "available": {
            "type": "boolean"
          },
          "unitId": {
            "type": "integer",
            "format": "int32"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          }
        },
        "additionalProperties": false
      },
      "Request": {
        "required": [
          "property",
          "responseBody",
          "responseSubject",
          "unit"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "integer",
            "format": "int32"
          },
          "property": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "unit": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time"
          },
          "toDate": {
            "type": "string",
            "format": "date-time"
          },
          "priceUponRequest": {
            "type": "number",
            "format": "double"
          },
          "confirmed": {
            "type": "boolean"
          },
          "processed": {
            "type": "boolean"
          },
          "sent": {
            "type": "boolean"
          },
          "numberOfPeople": {
            "type": "integer",
            "format": "int32"
          },
          "responseSubject": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "responseBody": {
            "type": "string"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          }
        },
        "additionalProperties": false
      },
      "Client": {
        "required": [
          "email",
          "name",
          "surname"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "surname": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "request": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Request"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}